+++ src/Rules/ArrayEnumRule.php
@@ -5,7 +5,7 @@ namespace Orisai\ObjectMapper\Rules;
 use Orisai\Exceptions\Logic\InvalidArgument;
 use Orisai\ObjectMapper\Args\Args;
 use Orisai\ObjectMapper\Args\ArgsChecker;
-use Orisai\ObjectMapper\Context\ArgsContext;
+use Orisai\ObjectMapper\Context\ArgsFieldContext;
 use Orisai\ObjectMapper\Context\FieldContext;
 use Orisai\ObjectMapper\Context\TypeContext;
 use Orisai\ObjectMapper\Exception\ValueDoesNotMatch;
@@ -29,7 +29,7 @@ final class ArrayEnumRule implements Rule
 		UseKeys = 'useKeys',
 		AllowUnknown = 'allowUnknown';
 
-	public function resolveArgs(array $args, ArgsContext $context): ArrayEnumArgs
+	public function resolveArgs(array $args, ArgsFieldContext $context): ArrayEnumArgs
 	{
 		$checker = new ArgsChecker($args, self::class);
 		$checker->checkAllowedArgs([self::Cases, self::UseKeys, self::AllowUnknown]);