+++ src/Rules/BoolRule.php
@@ -4,7 +4,7 @@ namespace Orisai\ObjectMapper\Rules;
 
 use Orisai\ObjectMapper\Args\Args;
 use Orisai\ObjectMapper\Args\ArgsChecker;
-use Orisai\ObjectMapper\Context\ArgsContext;
+use Orisai\ObjectMapper\Context\ArgsFieldContext;
 use Orisai\ObjectMapper\Context\FieldContext;
 use Orisai\ObjectMapper\Context\TypeContext;
 use Orisai\ObjectMapper\Exception\ValueDoesNotMatch;
@@ -30,7 +30,7 @@ final class BoolRule implements Rule
 		0 => false,
 	];
 
-	public function resolveArgs(array $args, ArgsContext $context): BoolArgs
+	public function resolveArgs(array $args, ArgsFieldContext $context): BoolArgs
 	{
 		$checker = new ArgsChecker($args, self::class);
 		$checker->checkAllowedArgs([self::CastBoolLike]);