+++ src/Rules/FloatRule.php
@@ -5,7 +5,7 @@ namespace Orisai\ObjectMapper\Rules;
 use Orisai\Exceptions\Logic\InvalidArgument;
 use Orisai\ObjectMapper\Args\Args;
 use Orisai\ObjectMapper\Args\ArgsChecker;
-use Orisai\ObjectMapper\Context\ArgsContext;
+use Orisai\ObjectMapper\Context\ArgsFieldContext;
 use Orisai\ObjectMapper\Context\FieldContext;
 use Orisai\ObjectMapper\Context\TypeContext;
 use Orisai\ObjectMapper\Exception\ValueDoesNotMatch;
@@ -30,7 +30,7 @@ final class FloatRule implements Rule
 		Unsigned = 'unsigned',
 		CastNumericString = 'castNumericString';
 
-	public function resolveArgs(array $args, ArgsContext $context): FloatArgs
+	public function resolveArgs(array $args, ArgsFieldContext $context): FloatArgs
 	{
 		$checker = new ArgsChecker($args, self::class);
 		$checker->checkAllowedArgs([self::Min, self::Max, self::Unsigned, self::CastNumericString]);