+++ src/Rules/InstanceOfRule.php
@@ -5,7 +5,7 @@ namespace Orisai\ObjectMapper\Rules;
 use Orisai\Exceptions\Logic\InvalidArgument;
 use Orisai\ObjectMapper\Args\Args;
 use Orisai\ObjectMapper\Args\ArgsChecker;
-use Orisai\ObjectMapper\Context\ArgsContext;
+use Orisai\ObjectMapper\Context\ArgsFieldContext;
 use Orisai\ObjectMapper\Context\FieldContext;
 use Orisai\ObjectMapper\Context\TypeContext;
 use Orisai\ObjectMapper\Exception\ValueDoesNotMatch;
@@ -23,7 +23,7 @@ final class InstanceOfRule implements Rule
 
 	private const Type = 'type';
 
-	public function resolveArgs(array $args, ArgsContext $context): InstanceOfArgs
+	public function resolveArgs(array $args, ArgsFieldContext $context): InstanceOfArgs
 	{
 		$checker = new ArgsChecker($args, self::class);
 		$checker->checkAllowedArgs([self::Type]);