+++ src/Rules/ListOfRule.php
@@ -4,7 +4,7 @@ namespace Orisai\ObjectMapper\Rules;
 
 use Orisai\ObjectMapper\Args\Args;
 use Orisai\ObjectMapper\Args\ArgsChecker;
-use Orisai\ObjectMapper\Context\ArgsContext;
+use Orisai\ObjectMapper\Context\ArgsFieldContext;
 use Orisai\ObjectMapper\Context\FieldContext;
 use Orisai\ObjectMapper\Context\TypeContext;
 use Orisai\ObjectMapper\Exception\InvalidData;
@@ -27,7 +27,7 @@ final class ListOfRule extends MultiValueRule
 
 	private const Continuous = 'continuous';
 
-	public function resolveArgs(array $args, ArgsContext $context): MultiValueArgs
+	public function resolveArgs(array $args, ArgsFieldContext $context): MultiValueArgs
 	{
 		$checker = new ArgsChecker($args, self::class);
 		$checker->checkAllowedArgs([self::ItemRule, self::MinItems, self::MaxItems, self::MergeDefaults]);