+++ src/Rules/MultiValueEfficientRuleAdapter.php
@@ -4,7 +4,7 @@ namespace Orisai\ObjectMapper\Rules;
 
 use Orisai\Exceptions\Logic\NotImplemented;
 use Orisai\ObjectMapper\Args\Args;
-use Orisai\ObjectMapper\Context\ArgsContext;
+use Orisai\ObjectMapper\Context\ArgsFieldContext;
 use Orisai\ObjectMapper\Context\FieldContext;
 use Orisai\ObjectMapper\Context\TypeContext;
 use Orisai\ObjectMapper\Types\Type;
@@ -26,7 +26,7 @@ final class MultiValueEfficientRuleAdapter implements MultiValueEfficientRule
 		$this->rule = $rule;
 	}
 
-	public function resolveArgs(array $args, ArgsContext $context): Args
+	public function resolveArgs(array $args, ArgsFieldContext $context): Args
 	{
 		$this->throwNotImplemented(__FUNCTION__);
 	}