+++ src/Rules/NullRule.php
@@ -4,7 +4,7 @@ namespace Orisai\ObjectMapper\Rules;
 
 use Orisai\ObjectMapper\Args\Args;
 use Orisai\ObjectMapper\Args\ArgsChecker;
-use Orisai\ObjectMapper\Context\ArgsContext;
+use Orisai\ObjectMapper\Context\ArgsFieldContext;
 use Orisai\ObjectMapper\Context\FieldContext;
 use Orisai\ObjectMapper\Context\TypeContext;
 use Orisai\ObjectMapper\Exception\ValueDoesNotMatch;
@@ -21,7 +21,7 @@ final class NullRule implements Rule
 
 	private const CastEmptyString = 'castEmptyString';
 
-	public function resolveArgs(array $args, ArgsContext $context): NullArgs
+	public function resolveArgs(array $args, ArgsFieldContext $context): NullArgs
 	{
 		$checker = new ArgsChecker($args, self::class);
 		$checker->checkAllowedArgs([self::CastEmptyString]);